user signup validations working
